Add pruning option to file.keyvalue
This implements a new optional argument "prune" in the file.keyvalue
state function, which, if set to True, deletes lines not configured
using the "key_values" or "key"/"value" arguments as part of the
execution. Useful for ensuring a file only contains the desired
key/value pairs and no additional data.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>